requirements.txt
streamlit==1.32.2
google-generativeai==0.8.2
pydantic==1.10.15
PyPDF2==3.0.1
python-dotenv==1.0.1
